chainiq-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.40.0
pandas==2.2.3
pydantic==2.9.2
streamlit==1.39.0
plotly==5.24.1
sse-starlette==2.1.3
httpx==0.27.2
python-dotenv==1.0.1
